fastapi/tasks-api/requirements.txt
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.13
python-decouple==3.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
